Add a base /etc/group file
Providing this file fixes a problem with the base-passwd package, which is
essential but provides passwd and group files only after being configured.
Package base-files runs a command in its postinst which depends on the presence
of a group file. It seems within base-files rights to depend on such without
declaring a dependency on base-passwd, since base-passwd is marked essential
and thus must provide its functionality without being configured. During an
install, if base-files gets configured before base-passwd, it fails for lack of
the group file.

One way to fix this might be to fix the base-passwd package. Another might be
to control the order in which base-files and base-passwd gets installed. Here
we decide to just write a group file from the present package, to satisfy
base-files need. We're already writing a base passwd file, and writing a group
file falls within the same scope.