Symlink /etc/mtab --> /proc/self/mounts
This is now the standard. Leaving it missing sometimes causes a a legacy
program to create it as a regular file, causing problems with, e.g.,
mount namespaces.

On a system we're upgrading, /etc/mtab may have already been created by
such a legacy program, so we delete it in our preinst.